backend/requirements.txt
# --- Core Backend ---
fastapi
uvicorn
supabase
python-dotenv
requests
pydantic

# --- AI & Machine Learning ---
sentence-transformers
numpy
scikit-learn